safari-600.7-branch/LayoutTests/java/lc3/ConvertNumber/number-008-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion Preferred argument conversion:  undefined

FAIL TEST_CLASS.ambiguous(1) should be BYTE. Was BYTE.
PASS successfullyParsed is true

TEST COMPLETE
 
